House imsdb parsing logic solely in the class
