+++ lib/utility/imsdb_parser.rb
@@ -5,6 +5,25 @@ module Utility
     SCRIPT_LINK_CSS_PATH='table td p a'.freeze
     LANDING_PAGE_CSS_PATH='.script-details td a'.freeze
 
+    def scrape_initial_movie_data
+      movie_links = all_movie_page_links
+
+      movie_links.each do |node|
+        create_movie_from_node(node)
+      end
+    end
+
+    private
+
+    def create_movie_from_node(node)
+      attributes = extract_movie_data_from_node(node)
+      # TODO: Create logic to find the writers, year, etc
+      Movie.create({
+        title: attributes[:title],
+        url: attributes[:url]
+      })
+    end
+
     def all_movie_page_links
       extract_tags_from_css_at_url(SCRIPT_LINK_CSS_PATH, ALL_SCRIPT_URL)
     end
@@ -16,7 +35,6 @@ module Utility
       }
     end
 
-    private
 
     def extract_title_from_node(node)
       attributes = node.attributes
@@ -42,4 +60,8 @@ module Utility
       "#{IMSBD_BASE_URL}#{path}"
     end
   end
+end
+
+def self.imsdb
+  Utility::ImsdbParser.new
 end
